requirements.txt
pandas
requests
psycopg2-binary
sqlalchemy
tomli
gql[aiohttp]
progressbar2